+++ sections/Carousel.tsx
@@ -1,3 +1,3 @@
-export { default } from "$store/components/ui/Carousel.tsx";
-export type { Image, Props } from "$store/components/ui/Carousel.tsx";
+export { default } from "$store/components/ui/BannerCarousel.tsx";
+export type { Image, Props } from "$store/components/ui/BannerCarousel.tsx";
 export type { Image as LiveImage } from "$live/std/ui/types/Image.ts";
